backend/requirements.txt
flask==2.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
pyjwt==2.8.0
sqlalchemy==2.0.19
flask-sqlalchemy==3.0.3
cryptography==41.0.2